Non-JSON serde buffer overflow: add fuzz target
Cardinality estimator makes assumptions about the values it receives when deserializing which are upheld by _`serde_json`_ but not `serde` in general, leading to buffer overflows and other problems when used with other serde de/serializers.

I had the wildest crashes happening in a project, which I traced to being at least _related_ to cardinality-estimator, since they stopped when I shimmed it out. In narrowing down a smaller reproducible example, I found that using bincode with this library was sufficient to get buffer overflows: see https://github.com/uniphil/whose-overflow-is-it-anyway

This PR adds a fuzz target for `postcard`, whose API is close enough to `serde_json` for the change from the serde fuzz target to be a single line. It also adds postcard calls to the normal serde tests, though I wasn't able to specifically write a failing regression test case for it (I'm sure it's possible, just had to stop somewhere).

To be clear: this change only illustrates the issue and does not include a fix. Seems like fixing will require either

- fixing the deserialization logic to assume less about the inputs it receives (array alignment and length?), OR
- documenting explicitly that the `serde` support is **only** safe to work with `serde_json`.

Thanks @DavidBuchanan314 for help narrowing the source of the crashes <3