refactor down to Sketch + Element to insert
this... kind of comes full circle to pushing the secret state back out to the caller. i think that's fine.

the api is more annoying now, but whatever. could make it fairly nice again via a few feature flags to just use wyhash again on insert with a special method.